refactor: change utils folder name to shared
